Better support for surrogate pairs and ligatures on Windows (GTK3)

See merge request GNOME/gtk!488

+++ gtk/gtkimcontextsimpleprivate.h
@@ -26,11 +26,11 @@ G_BEGIN_DECLS
 
 extern const GtkComposeTableCompact gtk_compose_table_compact;
 
-gboolean gtk_check_algorithmically (const guint16                *compose_buffer,
+gboolean gtk_check_algorithmically (const guint32                *compose_buffer,
                                     gint                          n_compose,
                                     gunichar                     *output);
 gboolean gtk_check_compact_table   (const GtkComposeTableCompact *table,
-                                    guint16                      *compose_buffer,
+                                    guint32                      *compose_buffer,
                                     gint                          n_compose,
                                     gboolean                     *compose_finish,
                                     gboolean                     *compose_match,
